_01_introduction/attempts/Stage5.java
// Cannot use Math library functions for any task

package practicePackage._01_introduction.attempts;

public class Stage5 {
	/**
	 * 
	 * @param data
	 * @return true if all items of the array are positive (more than 0), false otherwise.
	 * return false if the array is null.
	 */
	public static boolean allPositives(int[] data) {
		if (data == NULL) {
			return false;
		}
		for (int i = 0; i <data.length;i++) {
			if (data[i] > 0) {
				
			}
		}
		return false; //to be completed
	}

	/**
	 * 
	 * @param data
	 * @return true if there exists at least one positive (more than 0) value in the array, false otherwise.
	 * return false if the array is null.
	 */
	public static boolean containsPositive(int[] data) {
		return false; //to be completed
	}

	/**
	 * 
	 * @param a
	 * @param b
	 * @return true if a and b have the same items in the same order,
	 * false otherwise 
	 * return false if either of array passed is null
	 */
	public static boolean identical(int[] a, int[] b) {
		return false; //to be completed
	}

	/**
	 * 
	 * @param data
	 * @param target
	 * @param endIndex
	 * @return true if target exists anywhere up to 
	 * index endIndex (inclusive) in the array passed, false otherwise.
	 * return false if array is null or endIndex is invalid 
	 * or target doesn't exist in the array up to index endIndex.	 
	 */
	public static boolean containsUptoIndex(int[] data, int target, int endIndex) {
		return false; //to be completed
	}

	/**
	 * 
	 * @param data
	 * @return an array containing the numbers of occurrences of the items in data.
	 * For example, if data = {10,20,10,30,10,20,30,10,10,30,30}, return {5, 2, 4}
	 * as there are three distinct items (10, 20, 30) and,
	 * the first "first-of-its-kind" item (10) occurs 5 times
	 * the second "first-of-its-kind" item (20) occurs 2 times
	 * the third "first-of-its-kind" item (30) occurs 4 times
	 *  
	 * return null if the array is null.
	 */
	public static int[] getFrequencyTable(int[] data) {
		return new int[] {}; //to be completed
	}
}
